requirements.txt
dnspython==2.6.1
pymongo==4.6.3
python-dotenv==1.0.1
schedule==1.2.1
